Remediation 2026 Chunk 4: decompose the god SessionsController
Split SessionsController's 17 endpoints along service boundaries into four focused
controllers, all sharing the api/sessions route prefix so every URL, method,
request/response shape, and status code is unchanged (zero breaking changes).

- SessionsController (slimmed): core session events via IHandleSessionEvent
  ({date}/all, pastdue, POST, PUT {id}, patient/{id}/discovery). Dropped the
  unused logger and the 3 services it no longer needs.
- BookingController (new): full booking lifecycle via IBookingService
  (statuses, confirm/cancel/complete/noshow/checkin/start-therapy, unconfirmed,
  upcoming, {id}/confirmations).
- ScheduleController (new): {date}/schedule-matrix via IScheduleMatrixService.
- SessionPaymentsController (new): {id}/payments via IPaymentRecordService.
- Each controller uses explicit [Route("api/sessions")] (no [controller] token);
  verified no action-template collisions. Each injects only the one service it
  needs — the basis for cleaner per-boundary claims policies later.
- Extracted the shared date helper to Web/Common/RouteDates.ParseOrToday.
- Error handling unchanged: exceptions bubble to the 3A GlobalExceptionHandler.

Tests: slimmed SessionsControllerTests constructor (session-event tests unchanged);
added init/DI smoke tests for the three new controllers. Full suite green (199).
Route-preservation curl matrix: docs/sessions-split-4-verification.md.
