+++ src/Web/Controllers/BookingController.cs
@@ -1,0 +1,131 @@
+using Microsoft.AspNetCore.Mvc;
+using Neurocorp.Api.Core.BusinessObjects.Lookups;
+using Neurocorp.Api.Core.BusinessObjects.Sessions;
+using Neurocorp.Api.Core.Interfaces.Services;
+
+namespace Neurocorp.Api.Web.Controllers;
+
+// Appointment booking lifecycle, split out of SessionsController (Chunk 4). Keeps the api/sessions
+// route prefix so all URLs are unchanged. Errors bubble to the global ProblemDetails handler (3A).
+[ApiController]
+[Route("api/sessions")]
+public class BookingController : ControllerBase
+{
+    private readonly IBookingService _bookingService;
+
+    public BookingController(IBookingService bookingService)
+    {
+        _bookingService = bookingService;
+    }
+
+    [HttpGet("statuses")]
+    [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(IEnumerable<LookupItem>))]
+    public async Task<IActionResult> GetStatuses()
+    {
+        var statuses = await _bookingService.GetStatusesAsync();
+        return Ok(statuses);
+    }
+
+    [HttpPut("{id}/confirm")]
+    [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(SessionEvent))]
+    [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status404NotFound)]
+    [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status400BadRequest)]
+    public async Task<IActionResult> ConfirmAppointment(int id, [FromBody] ConfirmationRequest request)
+    {
+        var result = await _bookingService.ConfirmAppointmentAsync(id, request);
+        return Ok(result);
+    }
+
+    [HttpPut("{id}/cancel")]
+    [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(SessionEvent))]
+    [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status400BadRequest)]
+    public async Task<IActionResult> CancelAppointment(int id, [FromBody] CancelRequest request)
+    {
+        var result = await _bookingService.CancelAppointmentAsync(id, request.Reason);
+        return Ok(result);
+    }
+
+    [HttpPut("{id}/complete")]
+    [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(SessionEvent))]
+    [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status400BadRequest)]
+    public async Task<IActionResult> CompleteAppointment(int id)
+    {
+        var result = await _bookingService.UpdateStatusAsync(id, 4); // Completed
+        return Ok(result);
+    }
+
+    [HttpPut("{id}/noshow")]
+    [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(SessionEvent))]
+    [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status400BadRequest)]
+    public async Task<IActionResult> NoShowAppointment(int id)
+    {
+        var result = await _bookingService.UpdateStatusAsync(id, 5); // NoShow
+        return Ok(result);
+    }
+
+    [HttpPut("{id}/checkin")]
+    [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(SessionEvent))]
+    [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status400BadRequest)]
+    public async Task<IActionResult> CheckInAppointment(int id)
+    {
+        var result = await _bookingService.UpdateStatusAsync(id, 6); // CheckedIn
+        return Ok(result);
+    }
+
+    [HttpPut("{id}/start-therapy")]
+    [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(SessionEvent))]
+    [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status400BadRequest)]
+    public async Task<IActionResult> StartTherapy(int id)
+    {
+        var result = await _bookingService.UpdateStatusAsync(id, 7); // InTherapy
+        return Ok(result);
+    }
+
+    [HttpGet("unconfirmed")]
+    [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(IEnumerable<SessionEvent>))]
+    public async Task<IActionResult> GetUnconfirmed([FromQuery] string? from = null, [FromQuery] string? to = null, [FromQuery] int days = 7)
+    {
+        DateOnly fromDate, toDate;
+        if (from != null || to != null)
+        {
+            fromDate = from != null ? DateOnly.Parse(from) : DateOnly.FromDateTime(DateTime.UtcNow);
+            toDate = to != null ? DateOnly.Parse(to) : fromDate.AddDays(7);
+        }
+        else
+        {
+            if (days > 30) days = 30;
+            fromDate = DateOnly.FromDateTime(DateTime.UtcNow);
+            toDate = fromDate.AddDays(days);
+        }
+        var sessions = await _bookingService.GetUnconfirmedAsync(fromDate, toDate);
+        return Ok(sessions);
+    }
+
+    [HttpGet("upcoming")]
+    [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(IEnumerable<SessionEvent>))]
+    public async Task<IActionResult> GetUpcoming([FromQuery] string? from = null, [FromQuery] string? to = null, [FromQuery] int days = 7)
+    {
+        DateOnly fromDate, toDate;
+        if (from != null || to != null)
+        {
+            fromDate = from != null ? DateOnly.Parse(from) : DateOnly.FromDateTime(DateTime.UtcNow);
+            toDate = to != null ? DateOnly.Parse(to) : fromDate.AddDays(7);
+        }
+        else
+        {
+            if (days > 30) days = 30;
+            fromDate = DateOnly.FromDateTime(DateTime.UtcNow);
+            toDate = fromDate.AddDays(days);
+        }
+        var sessions = await _bookingService.GetUpcomingAsync(fromDate, toDate);
+        return Ok(sessions);
+    }
+
+    [HttpGet("{id}/confirmations")]
+    [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(IEnumerable<ConfirmationRecord>))]
+    public async Task<IActionResult> GetConfirmations(int id)
+    {
+        var confirmations = await _bookingService.GetConfirmationsAsync(id);
+        return Ok(confirmations);
+    }
+}
